Nice changes to build.sbt that allow us to use maven to download libdependencies on demand.
We looked up all required dependencies in MongoActSynth on search.maven.org
and were able to find most of them. Those that we didnt find, we googled
one way or another and were able to get the dependencies resolved by
putting the right requirements in build.sbt.

What this means is that the repo does not need to contain any of the libraries.
sbt will pull them on demand, when you do sbt compile.