Incorporating metacyc into the system.
The subsystem being checked in here reads biopax-level3.owl files.
It is to be called using sbt "run metacyc" and that goes into act.installer.Main { "METACYC" } and then calls the main parser in MetaCyc

Main classes are:
- MetaCyc: queries for all owl files and processes them into a file -> OrganismComposition
- BioPaxFile: Moves data from biopax model level3 to our data structures.
- Resource: string id encapsulator
- OrganismComposition: Holds maps that translate ID ->
        {
          Entities: Protein, Complex, SmallMolecule, ProteinRnaRef (holds sequences), ChemicalStructure (holds CML/SMILES),
          Processes: Pathways, BiochemicalPathwayStep, Catalysis and Modulation, Conversion (BiochemicalReaction, Transport, TransportWithBiochemicalReaction),
          Annotations:
          References: Unifications (DB:id), Relationships (Unifs + reln type), Publications, Evidence, Provenance
        }

It takes about 1 second per level3 file, and there are 3528 files, so about 1 hour to process all.