Chemicals installed with Stereochemistry.
The main flag that controls whether stereochemistry is stripped or not is in src/main/java/act/client/CommandLineRun.java: _clearAllStereo = false; (false means keep stereochemistry) and used in consistentInChI, which is the main function called to bring all inchis into a consistent space.

Some hardcoded chemicals had to be moved back to their raw data form (over which consistentInChI is called and depending on the flag the right form is installed). Hardcoded chemicals were in BalanceEquations.java (balancing molecules h2o, co2 etc identified by inchi), MongoDB.java (manually markedreachables, and definitecofactors are identified by inchi. definitecofactors is just a list to ensure that when creating a tree we do not assign the cofactors as parents, and is not the authoritative source for cofactor identification. See Issue#3 on github and cofactors-expanded-issue3-on-github.xlsx for that.), CreateActTree.java (the clade ancestors are identified by inchi)

Also note that as of this checkin, initdb will add brenda, kegg, and metacyc data to the system. We should not add kegg to the final authoritative set because we cannot correlate the kegg reactions back to sequences properly.