initial code for SAR calculation in act/installer/sar/SAR.java
called using sbt "runMain com.act.reachables.initdb infer_sar Q84LM4" or just sbt "runMain com.act.reachables.initdb infer_sar" (for inferring all SAR of all enzymes). Right now just calculates the MCS of all substrates (that have diversity) accepted by an enzyme.

But this is not what an SAR constraint will really look like. In fact, it will be the opposite. It will be designed to "explain" each negative observation. For each negative observation we will find the positive observation with which it has the "biggest" MCS. Subtract the resulting positive from this negative observation will tell us what SHOULD NOT BE PRESENT; i.e., what causes loss of activity.

See the discussion on page 15 of the Oct 7th, 1000 molecules report to DARPA.

We could introduce a "soft constraint" which the MCS of all positive observations. If a new substrate contains 1) (soft constraint) this positive MCS substructure and 2) (hard constraint) NONE of the negative "explainers" -- then we can be assume the enzyme will accept this substrate (of course if it contains the reactive center as described by the ERO "lhs precondition").

The code here uses the MCS algorithm we have in Molecules/MCS.java (which in turn uses Molecules/SMILES.java and Molecules/MolGraph.java)