SplitProject working from ChemAxon, with some edge cases
Have SplitChem, SplitReaction, and SplitProjector working, in the sense
that they generate appropriate products for tartrate and menthol
examples.

Known issues:
1)  I haven't populated the inversion array, but it is probably rare
that this will result in errors
2)  Haven't dealt with the scenario where there are multiple ways the
reaction can apply to a substrate.  This needs to happen during parsing
of the SplitReaction by inspection.  It would manifest itself as having
multiple equivalent solutions during AAM, but I'm not sure how to
inform that.  The example is tartrate --> methyltartrate. The methyl
can go on either side of the molecule, and in so doing further breaks
the symmetry of the molecule.  The current implementation will not
generate the "wrong" product, but it will only show one of the
potential products.