With this commit, it is doing BRO + orphans
In the original description of ERO, CRO, BRO, we assumed that the
reaction was stoichiometrically balanced.  In the cleanup, these things
aren't balanced, so there are orphan atoms.  The previous commit just
identified the orphans.  Now it is truly calculating the BRO -- the
atoms whose bonds moved, but also includes the orphans.

I'm not sure yet if I want to do cleanup with this abstraction, or want
to get more specific.  What might make sense is to bring in other atoms
on carbons, or do full ERO.  This also ignores double bond / single
bond differences, and I should probably consider a bond as "changed" if
the bond order changes.