RO now also keeps atoms that change bond order
Also added hashing out.  Played with the MappingStyle of the
AutoMapper.  I think that more directly gives the atoms we want here,
but it has the same issue of not respecting mechanism in that answer.
The code can probably be simplified dramatically with the CHANGING
style; wouldn't have to index everything out and find what's changed.