Added a timer
A few (like 1 in 50,000) reactions get stuck.  It happens most on
SkeletonMapping, but also sometimes occurs in regular AutoMapping
(ROExtractor).  I suspect this is unavoidable, so I've added a 30
second timer to avoid it getting stuck.  Sometimes there is also a
database retrieval error that results in termination, but no idea how
to prevent that.