Refactoring and ERO
Did various refactoring, and changed the skeleton mapper and change
mapper to just be map methods, consolidated the pro methods into one in
OperatorExtractor. Then added ERO to that as well.

ERO has been tested on a very short list of things, and required
extensive adjustment of the logic to get right, but happy with it on
correctly pulling in aromatics, double bonds, the first sigma shell.