Figuring out downstream processing
It's still running the original EROs, but working on downstream
processing.

CompareROs has been changed herein to use the 2 in each or >= 10
reactions to keep an ERO.  As those are hcEROs, they are the most
expressive and thus a high rate of occurrence is the strongest filter
for keeping the ERO.  I think that is appropriate.  CompareROs then
returns all the indices of the reactions that do these filtered ROs.

CrawlAndHmERO executes CompareROs then iterates through all those
filtered reactions and indexes the HmERO in an OperatorHasher.

CompareROs2 currently reads the serialized OperatorHasher, ranks them,
then prints them out. I'll next need to change it so that it renders a
GUI.