Code that generates a directory of ROs and reaction data from mappings
This has already been run.  It inputs two directories of mapped
reactions and "simple reactions" generated at a previous step.  It
calculates the various ROs and aggregates similar things into similar
file paths.

The ReactionInterpretation is a representation of a concrete reaction
that includes the ROs that are calculated for it, the AAM, and the
cofactors.  They are generated in the terminal leaf directories
generated by CrawlAndHmERO.