Rough out of new MechanisticValidator
Noticed several problems with the old code, so cannibalizing the old
one and building the new one that will handle both scanning all ROs as
well as running just one.  It also includes inspection of "FAKE"
cofactors.  It also generates a Report with an error log and details of
what occurred.  I haven't added the RO projection steps, but after that
will delete the old one.  Currently it seems to be pulling out the
cofactors and logging correctly on the test rxn.